Remove specific dictionaries and make the wiktionary one the main one
