+++ packages/word-definitions/src/lib/index.ts
@@ -1,2 +1,3 @@
 export { default as normalizeDefinition } from './normalizeDefinition';
 export { default as request } from './request';
+export { default as unique } from './unique';
