+++ packages/word-definitions/src/lib/unique.ts
@@ -1,0 +1,3 @@
+const unique = <T>(array: T[]): T[] => Array.from(new Set(array));
+
+export default unique;
