+++ packages/word-definitions/src/types.ts
@@ -1,0 +1,19 @@
+export interface WiktionaryResponse {
+  batchcomplete: string;
+  warnings: {
+    extracts: {
+      '*': string;
+    };
+  };
+  query: {
+    pages: Record<
+      string,
+      {
+        pageid: number;
+        ns: number;
+        title: string;
+        extract: string;
+      }
+    >;
+  };
+}
